git-svn-id: file:///semplestb/final_repo@1915 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SEMplestNIOProtocol/src/semplest/server/protocol/adengine/ReportObject.java
@@ -29,7 +29,7 @@ public class ReportObject
 	private Long averageCPC;
 	private Integer qualityScore;
 	private String approvalStatus;
-	private Integer firstPageCPC;
+	private Long firstPageCPC;
 	private Date createdDate;
 	private Long microCost;
 	
@@ -117,11 +117,11 @@ public class ReportObject
 	{
 		this.approvalStatus = approvalStatus;
 	}
-	public Integer getFirstPageCPC()
+	public Long getFirstPageCPC()
 	{
 		return firstPageCPC;
 	}
-	public void setFirstPageCPC(Integer firstPageCPC)
+	public void setFirstPageCPC(Long firstPageCPC)
 	{
 		this.firstPageCPC = firstPageCPC;
 	}
